acabando api d listaReproduccion y contenidos

+++ laravel/routes/api.php
@@ -5,7 +5,6 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Api\ListasReproduccionController;
 use App\Http\Controllers\Api\TokenController;
 use App\Http\Controllers\Api\ContenidosController;
-
 /*
 |--------------------------------------------------------------------------
 | API Routes
@@ -21,11 +20,11 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::apiResource('ListasReproduccion', ListasReproduccionController::class);
 Route::get('/user', [TokenController::class, 'user'])->middleware('auth:sanctum');;
 Route::post('/register', [TokenController::class, 'register']);
 Route::post('/login', [TokenController::class, 'login']);
 Route::post('/logout', [TokenController::class, 'logout'])->middleware('auth:sanctum');
 Route::apiResource('contenidos', ContenidosController::class);
-Route::post('/contenidos/{contenido}/guardar', [ContenidosController::class, 'guardar']);
-Route::delete('/contenidos/{contenido}/guardar', [ContenidosController::class, 'quitarGuardados']);
+Route::post('/contenidos/{contenido}/guardar/{id_lista}', [ContenidosController::class, 'guardar'])->middleware('auth:sanctum');
+Route::delete('/contenidos/{contenido}/guardar/{id_lista}', [ContenidosController::class, 'quitarGuardados'])->middleware('auth:sanctum');
+Route::apiResource('listas_reproduccion', ListasReproduccionController::class);
